FOR_UPSTREAM [VPG]: Ungate Display Pipe A subsystem Control before accessing display register
On CHV, before accessing dp_aux_ch register or DPIO register, we need to make
sure "Display Pipe A subsystem Control", or else we will read a Junk data which
might fail display.

FIXME: The Punit in current hardware doesn't seem ready for this yet, so leave
it iffed out. As of now, PIPE A hardcode.

Change-Id: Ie8579c9ce3cda36cabccb5d0d5bdaeb8c2d97661
Signed-off-by: Deepak S <deepak.s@intel.com>